+++ src/test/java/likelion/demo/repository/MemberRepositoryTest.java
@@ -2,6 +2,7 @@ package likelion.demo.repository;
 
 import jakarta.transaction.Transactional;
 import likelion.demo.domain.Member;
+import likelion.demo.domain.Playlist;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -15,6 +16,9 @@ public class MemberRepositoryTest {
 
     @Autowired      // 스프링의 의존성 주입 -> MemberRepository 인스턴스를 자동으로 주입
     private MemberRepository memberRepository;
+
+    @Autowired
+    private PlaylistRepository playlistRepository;
 
     @Test       // 해당 메서드가 테스트 메서드임을 나타냄
     public void joinFailIfEmailExist() {
@@ -34,6 +38,33 @@ public class MemberRepositoryTest {
 
         assertThrows(DataIntegrityViolationException.class, () -> {
             memberRepository.save(member2);
+        });
+    }
+
+    @Test
+    public void makePlaylistFail() {
+        // 멤버를 만듦
+        Member member = Member.builder()
+                .name("Chaelin")
+                .email("chaelin@mutsa.com")
+                .build();
+        memberRepository.save(member);
+
+        // 플레이리스트 생성
+        Playlist playlist1 = Playlist.builder()
+                .name("My Playlist")
+                .owner(member)
+                .build();
+        playlistRepository.save(playlist1);
+
+        Playlist playlist2 = Playlist.builder()
+                .name("My Playlist") // 같은 이름의 플레이리스트 생성
+                .owner(member)
+                .build();
+
+        // 테스트
+        assertThrows(DataIntegrityViolationException.class, () -> {
+            playlistRepository.save(playlist2);
         });
     }
 }
